HAAR-2001: add hidden role and endpoint for adding hidden role for a service for IMS

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/externalusersapi/resource/UserRoleController.kt
@@ -19,6 +19,7 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
+import uk.gov.justice.digital.hmpps.externalusersapi.config.AuthenticationFacade
 import uk.gov.justice.digital.hmpps.externalusersapi.config.ErrorResponse
 import uk.gov.justice.digital.hmpps.externalusersapi.resource.data.UserRoleDto
 import uk.gov.justice.digital.hmpps.externalusersapi.service.UserRoleService
@@ -28,6 +29,7 @@ import java.util.UUID
 @Tag(name = "/users/{userId}/roles", description = "User Roles Controller")
 class UserRoleController(
   private val userRoleService: UserRoleService,
+  private val authenticationFacade: AuthenticationFacade,
 ) {
   companion object {
     private val log = LoggerFactory.getLogger(this::class.java)
@@ -88,7 +90,7 @@ class UserRoleController(
 
   @DeleteMapping("/users/{userId}/roles/{role}")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  @PreAuthorize("hasAnyRole('ROLE_MAINTAIN_OAUTH_USERS', 'ROLE_AUTH_GROUP_MANAGER')")
+  @PreAuthorize("hasAnyRole('ROLE_MAINTAIN_OAUTH_USERS', 'ROLE_AUTH_GROUP_MANAGER', 'ROLE_MAINTAIN_IMS_USERS')")
   @Operation(
     summary = "Remove role from user.",
     description = "Remove role from user.",
@@ -149,13 +151,18 @@ class UserRoleController(
     @PathVariable
     role: String,
   ) {
-    userRoleService.removeRoleByUserId(userId, role)
+    if (authenticationFacade.getAuthentication().authorities.map { it.authority }.any { it == "ROLE_MAINTAIN_IMS_USERS" }
+    ) {
+      userRoleService.serviceRemoveRoleByUserId(userId, role)
+    } else {
+      userRoleService.removeRoleByUserId(userId, role)
+    }
     log.info("Remove role succeeded for userId {} and role {}", userId, role)
   }
 
   @PostMapping("/users/{userId}/roles")
   @ResponseStatus(HttpStatus.NO_CONTENT)
-  @PreAuthorize("hasAnyRole('ROLE_MAINTAIN_OAUTH_USERS', 'ROLE_AUTH_GROUP_MANAGER')")
+  @PreAuthorize("hasAnyRole('ROLE_MAINTAIN_OAUTH_USERS', 'ROLE_AUTH_GROUP_MANAGER', 'ROLE_MAINTAIN_IMS_USERS')")
   @Operation(
     summary = "Add roles to user.",
     description = "Add role to user, post version taking multiple roles.  Requires role ROLE_MAINTAIN_OAUTH_USERS or ROLE_AUTH_GROUP_MANAGER",
@@ -227,7 +234,13 @@ class UserRoleController(
     @NotEmpty
     roles: List<String>,
   ) {
-    userRoleService.addRolesByUserId(userId, roles)
+    // separate on role
+    if (authenticationFacade.getAuthentication().authorities.map { it.authority }.any { it == "ROLE_MAINTAIN_IMS_USERS" }
+    ) {
+      userRoleService.serviceAddRolesByUserId(userId, roles)
+    } else {
+      userRoleService.addRolesByUserId(userId, roles)
+    }
     log.info("Add role succeeded for userId {} and roles {}", userId, roles.toString())
   }
